+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add API filters for ManyToMany fields using the `__in` filter
+- Fix the Swagger generator to include all the queryparams
 
 ### Changed